Enable the required functionality(!) to do with reply_id.
It appears that (a) messages were not getting a reply-id and (b)
StompHandler.listen() was returning the first message it saw and not
allowing the calling program to check the reply id.

StompHandler should not care about our custom headers or about The
Protocol, so I've made a change to listen() such that if the block
passed to yield returns false, the message is not consumed and listen()
goes around again for the next message. This should play well with any
existing code without change since most things that could be returning
from the block are truthy, so it's not *exactly* a breaking change.

NebRequest now uses this functionality to check for the right message
before returning it as the response.

Changed tests to fill the replyID so that they pass correctly.  Since
NebRequest *always* sets replyID, and thus the above behaviour should
*always* happen, no additional tests are needed.

--HG--
branch : dev