Fixes from Kevin test.
Change the logging, quite a bit.

Change StompHandler to use the Stomp gem's "connection" class, not
"client" -- a completely different API for some reason -- which allows
us to finally get rid of all that tedious thread code in favour of
something much simpler.  Which apparently did not work anyway.  And
removes the LocalJumpError issue.